==> 1-ReadableFormatCleaning.ipynb <==
### Task : Cleaning twitter data from file Climate_change.csv 

#Importing Dependencies 
import csv
import pandas as pd
import numpy as np
from datetime import datetime

#Importing the below block to display all outputs 
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

%%time
#Importing the csv file as individual lines hence the formatting was incorrect
with open("climate-change.csv", 'r', encoding="utf8") as csv_file: 
    csv_reader = csv.reader(csv_file, delimiter=',')
    df = pd.DataFrame([csv_reader])

#Displaying the DataFrame
df.head()

#Transpoding the df and changing the column name of ease of readability
Transposed_df = df.T
Transposed_df.rename(columns={0:'RawText'}, inplace=True)
Transposed_df.head()

#Deleting Rows which are empty rows which represents empty lines
Transposed_df = Transposed_df[Transposed_df['RawText'].map(lambda d: len(d)) > 0]
Transposed_df.head()
Transposed_df.shape

--------------------------------------------------------------------------------------------------------------------------
**Upon investagating the raw file we can see that if a tweet has commas or if the tweet was in multiple lines the file formatted considered it as a new column value or a new line respectively Keeping that in mind below are the steps to format it right**
    1. If a tweet was seperated into multiple columns for having a comma bring them together to make it a single string.
    2. If a tweet was sperated into multiple lines even those had to be combined into a single line.

#Storing Column Headers
Column_headers = Transposed_df.loc[0][0]
print(Column_headers)

#Deleting first header row
Transposed_df = Transposed_df.iloc[1: , :]
Transposed_df.reset_index(inplace=True, drop=True)

#Function to check if the first value in the list is a date and return date if true and return empty string if not 
def timecheck(value):
    try:
        if type(datetime.strptime(value[0],'%Y-%m-%d %H:%M:%S')) == type(now):
            return value[0]
        else:
            return ''
    except:
        return ''

%%time
# Applying the above function to check if the first value of the list in each line is a date 
# as that denotes the start of the line of a new tweet
Transposed_df['date'] = Transposed_df['RawText'].apply(lambda x: timecheck(x))
Tweet_Rows = Transposed_df[Transposed_df['date'] != ''].index.tolist() #Saving the index of starting rows of tweets for future use

#Investigating on the maximum columns the tweet has spread out due to commas in the tweet string
Transposed_df['list_len'] = Transposed_df['RawText'].apply(lambda x: len(x))
max(Transposed_df['list_len'])

Transposed_df.head(10)

Tweet_Rows[:10]
len(Tweet_Rows)

#Exploring a row which has list length more than 6
Transposed_df['RawText'][4]

#The above example shows if a tweet has commas then the string after the delimiter was considered as a new cell value
# Creating a new column which captures all these values into one single column
Individual_tweets =  Transposed_df.loc[Tweet_Rows]     #First creating a copy of this df for processing

print(Transposed_df.shape)
print(Individual_tweets.shape)

Individual_tweets.head()

#Creating Empty columns to input the spilled texts
Individual_tweets["nextcolTweets"] = ''
Individual_tweets["nextLineTweets"] = ''

%%time
# Concataniting the column spill of texts into one single string.
Individual_tweets['nextcolTweets'] = Individual_tweets['RawText'].apply(lambda x: ' '.join(x[6:len(x)]) if(len(x)>6) else '')

print("Number of tweet strings spilled to other columns : ", (Individual_tweets['nextcolTweets'].values != '').sum())

Individual_tweets.head() #Displaying the O/P

#Now that the Column spill is taken care of lets fix the next line spill
Transposed_df.head() #Displaying the main df

#Creating an array of row indexes and next index which will be used to concatnate rows data if spilled.
len(Tweet_Rows)
inc_Tweet_Rows = Tweet_Rows[1:]
inc_Tweet_Rows.append(Tweet_Rows[-1]) ## Appending the last element 
len(inc_Tweet_Rows)
Tweet_Rows[0:10]
inc_Tweet_Rows[0:10]

#Creating a list of 
# arrayvalue = np.array([Tweet_Rows,inc_Tweet_Rows]).T
#Individual_tweets["npArray"] = pd.Series(arrayvalue.tolist()) # This line doesnt work for last rows unsure why hence the approach in next lines

#Putting them into the dataframe column for lookup
Individual_tweets["StartRow"] = Tweet_Rows
Individual_tweets["EndRow"] = inc_Tweet_Rows

#Combining these two into an array for lookup
Individual_tweets["indexArray"] = Individual_tweets[["StartRow", "EndRow"]].values.tolist()
Individual_tweets.tail()

# Performing a lookup to move the values into the main df 
Transposed_df["indexArray"] = Transposed_df.index.map(Individual_tweets["indexArray"]) 
Transposed_df.head()
Transposed_df.tail()

# Creating a function which concatnated the strings in multiple lines into one and adding it to one cell 
def getnextlinetweets(pointer):
    return ' '.join([j for i in Transposed_df.RawText[pointer[0]+1:pointer[1]] for j in i])

%%time
# Performing the concat to captur the data on a new row.
Transposed_df["nextLineTweets"] = Transposed_df["indexArray"].apply(lambda x: getnextlinetweets(x) if(type(x) == list) else '')

#Look's up values in nextcolTweets to main df
Transposed_df["nextcolTweets"] = Transposed_df.index.map(Individual_tweets["nextcolTweets"])

Transposed_df.head()
Transposed_df.tail()

# Deleting all the extra lines and creating a copy of new df
PartiallyCleaned = Transposed_df[Transposed_df["date"] != '']
PartiallyCleaned.head(5)

#Extracting useful columns only
Column_headers

#Moving the data from main RawText into individual columns
PartiallyCleaned["user"] = PartiallyCleaned.RawText.apply(lambda x: x[1])
PartiallyCleaned["is_retweet"] = PartiallyCleaned.RawText.apply(lambda x: x[2])
PartiallyCleaned["is_quote"] = PartiallyCleaned.RawText.apply(lambda x: x[3])
PartiallyCleaned["text"] = PartiallyCleaned.RawText.apply(lambda x: x[4])
PartiallyCleaned["quoted_text"] = PartiallyCleaned.RawText.apply(lambda x: x[5] if (len(x) > 5) else '')

#Deleting unwanted rows
PartiallyCleaned = PartiallyCleaned[['date', 'user', 'is_retweet', 'is_quote', 'text', 'quoted_text','nextLineTweets','nextcolTweets']]
PartiallyCleaned.head() 

PartiallyCleaned.tail() 

#Saving The file for NLP cleaning
PartiallyCleaned.reset_index(inplace=True, drop=True)
PartiallyCleaned.to_csv("Climate_change_ImportableFormat.csv", index=False)

#Reading the file back to test import
formatted = pd.read_csv("Climate_change_ImportableFormat.csv", keep_default_na=False,  dtype=object)
formatted.head()
