==> 6-BiLSTM_Char_CNN_non_GloVe.ipynb <==
#Importing dependencies for EDA
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import re
import json

import spacy
from spacy import displacy

# from validation import compute_f1
from keras.models import Model
from keras.layers import TimeDistributed,Conv1D,Dense,Embedding,Input,Dropout,LSTM,Bidirectional,MaxPooling1D,Flatten,concatenate
# from prepro import readfile,createBatches,createMatrices,iterate_minibatches,addCharInformatioin,padding
from tensorflow.keras.utils import Progbar
from keras.preprocessing.sequence import pad_sequences
from keras.initializers import RandomUniform

#Importing the below block to display all outputs 
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

#Reading the training file
df = pd.read_csv("/content/drive/MyDrive/NLP/ner_datasetreference.csv", encoding='latin')
df = df.fillna(method='ffill')
df["Sentence #"] = df["Sentence #"].apply(lambda s: s[9:]).astype("int32")
df["Word"] = df["Word"].apply(lambda x: re.sub(r'http\S+', ' ',x))

#Creating a dictionary to replace these latin characters 
replacement_dict = {'ë':'e','ü':'u',"\xa0":' ', 'é':'e', '\x93':' ','\x91':' ','\x97':' ','\x85':' ','\x94':' ','ö':'o' ,'°':' ', 
                   '\x92':' ','\x96':' '}

def cleanunicode(uncleanstring):
    return re.sub(r'[\xa0|é|ë|ü|\x93|\x91|\x97|\x85|\x94|ö|°|\x92|\x96]', lambda m: replacement_dict.get(m.group()), uncleanstring)

df['Word'] = df['Word'].apply(lambda x : cleanunicode(x))

# Creating a list of all unique words and unique tags
all_words = list(set(df["Word"].values))
all_tags = list(set(df["Tag"].values))
print("Number of unique words: {}".format(len(all_words)))
print("Number of unique tags : {}".format(len(all_tags)))

# Creating a dictionary of of unique words and assigning a unique index number similarly to NER tags as well 
word2Idx = {word: idx + 2 for idx, word in enumerate(all_words)}
word2Idx["UNKNOWN_WORD"]=0
word2Idx["PADDING"]= 1

print(dict(list(word2Idx.items())[0:5])) ## Dislaying 5 samples


idx2Word = {v: k for k, v in word2Idx.items()}

with open('/content/drive/MyDrive/NLP/idx2Word_nonGlove.json', 'w') as fp:
    json.dump(idx2Word, fp)

len(word2Idx)

# Word and tag is combined in a list for each sentence another list is consuructed
df['combined']= df[['Word','Tag']].values.tolist()
df = df.groupby(['Sentence #'])['combined'].agg(lambda x: list(x))

df.head()

trainSentences = df.tolist()

len(trainSentences)

from sklearn.model_selection import train_test_split

X_train, X_test = train_test_split(trainSentences, test_size=0.2, random_state=123)

len(X_train)
len(X_test)

X_train[0]

#This function creates a list of characters in a word
def characterSplit(Sentences):
    for i,sentence in enumerate(Sentences):
        for j,data in enumerate(sentence):
            chars = [c for c in data[0]]
            Sentences[i][j] = [data[0],chars,data[1]]
    return Sentences

X_train = characterSplit(X_train)
X_test = characterSplit(X_test)

X_train[0]

#This loops create a set of unique tags and a dictionary of unique words
labelSet = set()
words = {}

for dataset in [X_train, X_test]:
    for sentence in dataset:
        for token,char,label in sentence:
            labelSet.add(label)
            words[token.lower()] = True

len(labelSet)
len(words)

# :: Create a mapping for the labels ::
label2Idx = {}
for label in labelSet:
    label2Idx[label] = len(label2Idx)

idx2Label = {v: k for k, v in label2Idx.items()}

with open('/content/drive/MyDrive/NLP/idx2Label_nonGlove.json', 'w') as fp:
    json.dump(idx2Label, fp)

# :: Hard coded case lookup ::
case2Idx = {'numeric': 0, 'allLower':1, 'allUpper':2, 'initialUpper':3, 'other':4, 'mainly_numeric':5, 'contains_digit': 6, 'PADDING_TOKEN':7}
caseEmbeddings = np.identity(len(case2Idx), dtype='float32')

case2Idx
caseEmbeddings

#simarlrly creating a character index dictionary of all possible english characters
char2Idx = {"PADDING":0, "UNKNOWN":1}
for c in " 0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ.,-_()[]{}!?:;#'\"/\\%$`&=*+@^~|":
    char2Idx[c] = len(char2Idx)
len(char2Idx)
char2Idx

with open('/content/drive/MyDrive/NLP/char2Idx_nonGlove.json', 'w') as fp:
    json.dump(char2Idx, fp)

def getCasing(word, caseLookup):   
    casing = 'other'
    
    numDigits = 0
    for char in word:
        if char.isdigit():
            numDigits += 1
            
    digitFraction = numDigits / float(len(word))
    
    if word.isdigit(): #Is a digit
        casing = 'numeric'
    elif digitFraction > 0.5:
        casing = 'mainly_numeric'
    elif word.islower(): #All lower case
        casing = 'allLower'
    elif word.isupper(): #All upper case
        casing = 'allUpper'
    elif word[0].isupper(): #is a title, initial char upper, then all lower
        casing = 'initialUpper'
    elif numDigits > 0:
        casing = 'contains_digit'
    
   
    return caseLookup[casing]

def createMatrices(document, word2Idx, label2Idx, case2Idx,char2Idx):
    unknownIdx = word2Idx['UNKNOWN_WORD']
    paddingIdx = word2Idx['PADDING']    
        
    dataset = []
    
    wordCount = 0
    unknownWordCount = 0
    
    for document in document:
        wordIndices = []    
        caseIndices = []
        charIndices = []
        labelIndices = []
        
        for word,char,label in document:  
            wordCount += 1
            if word in word2Idx:
                wordIdx = word2Idx[word]
            elif word.lower() in word2Idx:
                wordIdx = word2Idx[word.lower()]                 
            else:
                wordIdx = unknownIdx
                unknownWordCount += 1
            charIdx = []
            for x in char:
                charIdx.append(char2Idx[x])
            #Get the label and map to int            
            wordIndices.append(wordIdx)
            caseIndices.append(getCasing(word, case2Idx))
            charIndices.append(charIdx)
            labelIndices.append(label2Idx[label])
           
        dataset.append([wordIndices, caseIndices, charIndices, labelIndices]) 
        
    return dataset

key_lenth = [len(i) for i,k in word2Idx.items()]
keys = {len(i):i for i,k in word2Idx.items()}
keys[max(key_lenth)]

#Padding the characters to length of 28
def padding(document):
    maxlen = 28
    for sentence in document:
        char = sentence[2]
        for x in char:
            maxlen = max(maxlen,len(x))
    for i,sentence in enumerate(document):
        document[i][2] = pad_sequences(document[i][2],28,padding='post')
    return document

#Above three functions are called on the train and test data
train_set = padding(createMatrices(X_train,word2Idx,  label2Idx, case2Idx,char2Idx))
test_set = padding(createMatrices(X_test, word2Idx, label2Idx, case2Idx,char2Idx))

#This function O/P the dataset and also o/p a list which contain the no of batches which has the same length words in i/p
def createBatches(data):
    l = []
    for i in data:
        l.append(len(i[0]))
    l = set(l)
    batches = []
    batch_len = []
    z = 0
    for i in l:
        for batch in data:
            if len(batch[0]) == i:
                batches.append(batch)
                z += 1
        batch_len.append(z)
    return batches,batch_len

train_batch,train_batch_len = createBatches(train_set)
test_batch,test_batch_len = createBatches(test_set)

len(train_set)
len(train_batch_len)

train_set[0]

def iterate_minibatches(dataset,batch_len): 
    start = 0
    for i in batch_len:
        tokens = []
        caseing = []
        char = []
        labels = []
        data = dataset[start:i]
        start = i
        for dt in data:
            t,c,ch,l = dt
            l = np.expand_dims(l,-1)
            tokens.append(t)
            caseing.append(c)
            char.append(ch)
            labels.append(l)
        yield np.asarray(labels),np.asarray(tokens),np.asarray(caseing),np.asarray(char)

# Hyper parameters used base on the paper minor changes as 
# the dataset i used to train is very small compared to the dataset they used for training

# EPOCHS = 50               # paper: 80
# DROPOUT = 0.5             # paper: 0.68
# DROPOUT_RECURRENT = 0.25  # not specified in paper, 0.25 recommended
# LSTM_STATE_SIZE = 200     # paper: 275
# CONV_SIZE = 3             # paper: 3
# LEARNING_RATE = 0.0105    # paper 0.0105
# OPTIMIZER = Nadam()       # paper uses SGD(lr=self.learning_rate), Nadam() recommended for smaller dataset

#Word Embedding
words_input = Input(shape=(None,),dtype='int32',name='words_input')
words = Embedding(input_dim=len(word2Idx), output_dim=20)(words_input) #,  weights=[wordEmbeddings], trainable=False

#Case embedding of each words
casing_input = Input(shape=(None,), dtype='int32', name='casing_input')
casing = Embedding(output_dim=caseEmbeddings.shape[1], input_dim=caseEmbeddings.shape[0], weights=[caseEmbeddings], trainable=False)(casing_input)

#Character embedding using 1D CNN
character_input=Input(shape=(None,28,),name='char_input')
embed_char_out=TimeDistributed(Embedding(len(char2Idx),30,embeddings_initializer=RandomUniform(minval=-0.5, maxval=0.5)), name='char_embedding')(character_input)
dropout= Dropout(0.5)(embed_char_out)
conv1d_out= TimeDistributed(Conv1D(kernel_size=3, filters=30, padding='same',activation='tanh', strides=1))(dropout)
maxpool_out=TimeDistributed(MaxPooling1D(28))(conv1d_out)
char = TimeDistributed(Flatten())(maxpool_out)
char = Dropout(0.5)(char)

#Concatnating the embedded layer o/p together
output = concatenate([words, casing, char])
output = Bidirectional(LSTM(200, return_sequences=True, dropout=0.50, recurrent_dropout=0.25))(output)
output = TimeDistributed(Dense(len(label2Idx), activation='softmax'))(output)

model = Model(inputs=[words_input, casing_input, character_input], outputs=[output])

model.compile(loss='sparse_categorical_crossentropy', optimizer='nadam')
model.summary()

%%time
#Training the model on batches
epochs = 50
for epoch in range(epochs):    
    print("Epoch %d/%d"%(epoch,epochs))
    a = Progbar(len(train_batch_len))
    for i,batch in enumerate(iterate_minibatches(train_batch,train_batch_len)):
        labels, tokens, casing,char = batch       
        history = model.train_on_batch([tokens, casing, char], labels,return_dict=True,reset_metrics=False)
        a.update(i)
    a.update(i+1)
    print(' ')

history

model.save("BiLSTM_CNN_nonGolVe/model.h5")

def predict_dataset(dataset):
    sentences = []
    correctLabels = []
    predLabels = []
    b = Progbar(len(dataset))
    for i,data in enumerate(dataset):    
        tokens, casing,char, labels = data
        sentences.append(tokens)
        tokens = np.asarray([tokens])     
        casing = np.asarray([casing])
        char = np.asarray([char])
        pred = model.predict([tokens, casing,char], verbose=False)[0]   
        pred = pred.argmax(axis=-1) #Predict the classes 
        correctLabels.append(labels)
        predLabels.append(pred)
        b.update(i)
    b.update(i+1)
    return predLabels, correctLabels, sentences

predLabels, correctLabels, sentences = predict_dataset(test_batch) 

#Returns the labels of index
def pred2label(pred,diction):
    out = []
    for pred_i in pred:
        out_i = []
        for p in pred_i:
            out_i.append(str(diction[p]))
        out.append(out_i)
    return out

pred_labels = pred2label(predLabels,idx2Label)
true_labels = pred2label(correctLabels,idx2Label)
actual_words = pred2label(sentences,idx2Word)

from seqeval.metrics import precision_score, recall_score, f1_score, classification_report

print("F1-score: {:.1%}".format(f1_score(true_labels, pred_labels)))
print("precision_score: {:.1%}".format(precision_score(true_labels, pred_labels)))
print("recall_score: {:.1%}".format(recall_score(true_labels, pred_labels)))

#Creating a flat list to look at Classification report
test_pred = [ j for i in pred_labels for j in i]
true_lab = [ j for i in true_labels for j in i]

from sklearn.metrics import classification_report
print(classification_report(test_pred, true_lab))

print("{:15}{:5}\t{}\n".format("Word", "True", "Pred"))
print("-"*30)

i = np.random.randint(0, len(actual_words))

for (w, t, pred) in zip(actual_words[i], true_labels[i], pred_labels[i]):
    print("{:15}{}\t{}".format(w, t, pred))

model.save("/content/drive/MyDrive/NLP/HalfTrainedModels/Half_Trained_model_non_glove.h5")

#Training on full dataset
#Model layers 

#Word Embedding part
words_input_full = Input(shape=(None,),dtype='int32',name='words_input')
words_full = Embedding(input_dim=len(word2Idx), output_dim=20)(words_input_full) # weights=[wordEmbeddings], trainable=False)(words_input_full)

#Case embedding part of each word
casing_input_full = Input(shape=(None,), dtype='int32', name='casing_input')
casing_full = Embedding(output_dim=caseEmbeddings.shape[1], input_dim=caseEmbeddings.shape[0], weights=[caseEmbeddings], trainable=False)(casing_input_full)

#Character embedding using 1D CNN
character_input_full = Input(shape=(None,28,),name='char_input')
embed_char_out_full = TimeDistributed(Embedding(len(char2Idx),30,embeddings_initializer=RandomUniform(minval=-0.5, maxval=0.5)), name='char_embedding')(character_input_full)
dropout_full = Dropout(0.5)(embed_char_out_full)
conv1d_out_full = TimeDistributed(Conv1D(kernel_size=3, filters=30, padding='same',activation='tanh', strides=1))(dropout_full)
maxpool_out_full =TimeDistributed(MaxPooling1D(28))(conv1d_out_full)
char_full = TimeDistributed(Flatten())(maxpool_out_full)
char_full = Dropout(0.5)(char_full)

#Concatnating the embedded layer o/p together
output_full = concatenate([words_full, casing_full, char_full])
output_full = Bidirectional(LSTM(200, return_sequences=True, dropout=0.50, recurrent_dropout=0.25))(output_full)
output_full = TimeDistributed(Dense(len(label2Idx), activation='softmax'))(output_full)

#Specifying input and o/p of the model
Full_Trained_model_non_glove = Model(inputs=[words_input_full, casing_input_full, character_input_full], outputs=[output_full])

#Compliling the model and looking at model summary
Full_Trained_model_non_glove.compile(loss='sparse_categorical_crossentropy', optimizer='nadam')
Full_Trained_model_non_glove.summary()

trainSentences = characterSplit(trainSentences)
full_train_data = padding(createMatrices(trainSentences ,word2Idx,  label2Idx, case2Idx, char2Idx))
full_train_data_batch,full_train_data_batch_len = createBatches(full_train_data)

%%time
#Training the model on full data with batches and for 50 epochs 
epochs = 50
for epoch in range(epochs):    
    print("Epoch %d/%d"%(epoch,epochs))
    a = Progbar(len(full_train_data_batch_len))
    for i,batch in enumerate(iterate_minibatches(full_train_data_batch,full_train_data_batch_len)):
        labels, tokens, casing, char = batch       
        Full_Trained_model_non_glove.train_on_batch([tokens, casing, char], labels)
        a.update(i)
    a.update(i+1)
    print(' ')

Full_Trained_model_non_glove.save("/content/drive/MyDrive/NLP/TrainedModels/Full_Trained_model_non_glove.h5")

from tensorflow import keras
LoadedModel = keras.models.load_model("TrainedModels/Full_Trained_model_non_glove.h5")

#Reading the dictionaries back 
with open('idx2Word_Glove.json', 'r') as fp:
    idx2Word = json.load(fp)
    
with open('idx2Label_Glove.json', 'r') as fp:
    idx2Label = json.load(fp)

#Readig the cleaned dataset
ClimateChange_DF = pd.read_csv("/content/drive/MyDrive/NLP_Files/NLP/TaskDataset/Cleaned_English_tweets.csv")
#keeping only the tweets
ClimateChange_DF = ClimateChange_DF[["CleanedTweets"]]
ClimateChange_DF.head()
ClimateChange_DF.shape

#selecting first 1000 tweets for ease of processing in next steps
ClimateChange_DF_Sliced = ClimateChange_DF.iloc[:10000].copy()

ClimateChange_DF_Sliced["Word"] = ClimateChange_DF_Sliced["CleanedTweets"].apply(lambda x: [[i] for i in str(x).split()])
ClimateChange_DF_Sliced.head()

climate_change = ClimateChange_DF_Sliced.Word.tolist()

#Splits the word into list of chatracters
def FcharacterSplit(Sentences):
    for i,sentence in enumerate(Sentences):
        for j,data in enumerate(sentence):
            chars = [c for c in data[0]]
            Sentences[i][j] = [data[0],chars]
    return Sentences

predicting_sentence = FcharacterSplit(climate_change)

#Addiing a dummy tag to process easilly
for i in predicting_sentence:
    for j in i:
        j.append('O')

climateChange_pred_data = padding(createMatrices(predicting_sentence, word2Idx, label2Idx, case2Idx,char2Idx))

i = np.random.randint(0, len(climateChange_pred_data))
print(i)
for j, data in enumerate(climateChange_pred_data[i:i+1]):
    tokens, casing, char, labels = data
    token = np.asarray([tokens])     
    casing = np.asarray([casing])
    char = np.asarray([char])
    pred = Full_Trained_model_non_glove.predict([token, casing,char], verbose=False)[0]   
    pred = pred.argmax(axis=-1) #Predict the classes 

#Returns the labels of index
def pred2label(pred,diction):
    out = []
    for pred_i in pred:
        out.append(diction[pred_i])
    return out

pred_labels = pred2label(pred,idx2Label)
actual_words = pred2label(tokens,idx2Word)

print("{:15}{:15}{}".format("Original_Word", "Passed_Word", "Tag_Predicted"))
print("-"*30)

for (o, w, pred) in zip(ClimateChange_DF_Sliced.CleanedTweets[i].split(), actual_words, pred_labels):
    print("{:15}{:15}{}".format(o, w, pred))

#Using Spacy to cross check for entities
nlp = spacy.load('en_core_web_sm')
text = nlp(ClimateChange_DF_Sliced.CleanedTweets[i])
displacy.render(text, style = 'ent', jupyter=True)
