==> 7-Predicting_Named_Entity.ipynb <==
### This notebook predicts the Named Entities for a selected string from Climate-Change dataset

#Importing libraries
from tensorflow import keras
import pandas as pd
import numpy as np
import json
import itertools
from keras.preprocessing.sequence import pad_sequences

#Custom library
from predifinedfunctions import characterSplit, addDummy, predictedDisplay, predictTags, padding, createMatrices, getCasing
# For prediction validation
import spacy
from spacy import displacy

#Importing NER Recognition models

#BILSTM
BiLSTM_Final_model = keras.models.load_model('TrainedModels/BiLSTM_Final_model.h5')
with open('word2index.json', 'r') as fp:
    word2index_1 = json.load(fp)    
with open('tag2index.json', 'r') as fp:
    tag2index_1 = json.load(fp)

#BILSTM-Strict-Clean
BiLSTM_Final_model_strict_clean = keras.models.load_model('TrainedModels/BiLSTM_Final_model_Strict_Clean.h5')
with open('word2index_strict_cleaned.json', 'r') as fp:
    word2index_2 = json.load(fp)    
with open('tag2index_strict_cleaned.json', 'r') as fp:
    tag2index_2 = json.load(fp)

#Bi-LSTM_CNN_GloVe    
BiLSTM_CNN_model_glove = keras.models.load_model('TrainedModels/Full_Trained_model_glove.h5')
with open('idx2Word_Glove.json', 'r') as fp:
    idx2Word_Glove = json.load(fp)    
with open('idx2Label_Glove.json', 'r') as fp:
    idx2Label_Glove = json.load(fp)

#Bi-LSTM_CNN_nonGloVe  
BiLSTM_CNN_model_non_glove = keras.models.load_model('TrainedModels/Full_Trained_model_non_glove.h5')
with open('idx2Word_nonGlove.json', 'r') as fp:
    idx2Word_nonGlove = json.load(fp)    
with open('idx2Label_nonGlove.json', 'r') as fp:
    idx2Label_nonGlove = json.load(fp)
    
#Importing additional dependencies    
with open('case2Idx.json', 'r') as fp:
    case2Idx = json.load(fp)    
with open('char2Idx.json', 'r') as fp:
    char2Idx = json.load(fp)
    

#Readig the Climate-Change Dataset and selecting top 10k tweets
ClimateChange_DF = pd.read_csv("Cleaned_English_tweets.csv")
ClimateChange_DF = ClimateChange_DF[["CleanedTweets"]].iloc[:10000].copy()
ClimateChange_DF.head()
ClimateChange_DF.shape

# Creating a list of word index with padding for the first two models
ClimateChange_DF["1Word"] = ClimateChange_DF["CleanedTweets"].apply(lambda x: str(x).split())
ClimateChange_DF["2Word"] = ClimateChange_DF["CleanedTweets"].apply(lambda x: str(x).lower().split())
ClimateChange_DF['1Word_Index'] = ClimateChange_DF['1Word'].apply(lambda x: [word2index_1[s] if s in word2index_1 else word2index_1['UNKNOWN_WORD'] for s in x])
ClimateChange_DF['2Word_Index'] = ClimateChange_DF['2Word'].apply(lambda x: [word2index_2[s] if s in word2index_2 else word2index_2['UNKNOWN_WORD'] for s in x])
ClimateChange_DF["1Padded_Word_Index"] = ClimateChange_DF["1Word_Index"].apply(lambda x: x + [word2index_1["PADDING"]] * (40 - len(x)) if (len(x) <=40) else x[:40])
ClimateChange_DF["2Padded_Word_Index"] = ClimateChange_DF["2Word_Index"].apply(lambda x: x + [word2index_2["PADDING"]] * (40 - len(x)) if (len(x) <=40) else x[:40])
ClimateChange_DF["3Word"] = ClimateChange_DF["CleanedTweets"].apply(lambda x: [[i] for i in str(x).split()])

ClimateChange_DF.head()

#Creating an input list for Glove model
model_input = ClimateChange_DF["3Word"].tolist()

#Splitting Characters
predicting_sentence = characterSplit(model_input)
#Preprocessing data by adding dummy tags           
predicting_sentence = addDummy(predicting_sentence)

#Creating lookup dictionaries for tags and words
index2Tag_1 = {value: key for key, value in tag2index_1.items()}
index2Tag_2 = {value: key for key, value in tag2index_2.items()}
word2Idx_Glove = {value: int(key) for key, value in idx2Word_Glove.items()}
word2Idx_nonGlove = {value: int(key) for key, value in idx2Word_nonGlove.items()}

#Instantiating model inputs
model_1_input = ClimateChange_DF["1Padded_Word_Index"]
model_2_input = ClimateChange_DF["2Padded_Word_Index"]
model_3_input = padding(createMatrices(predicting_sentence, word2Idx_Glove, case2Idx, char2Idx), 52)
model_4_input = padding(createMatrices(predicting_sentence, word2Idx_nonGlove, case2Idx, char2Idx), 28)

# Predicting NER using all 4 models and displaying the words with their tags

index = np.random.randint(0, 10000)
print(index)
#Model Prediction
model1_pred, model2_pred , model3_pred, model4_pred = predictTags(index, 
                                                                  BiLSTM_Final_model, model_1_input,
                                                                  BiLSTM_Final_model_strict_clean, model_2_input,
                                                                  BiLSTM_CNN_model_glove, model_3_input,
                                                                  BiLSTM_CNN_model_non_glove, model_4_input)
#Predicted Entity display                                                                 
predictedDisplay(ClimateChange_DF, index, model1_pred, index2Tag_1, model2_pred,index2Tag_2, model3_pred,idx2Label_Glove,
                 model4_pred, idx2Label_nonGlove)

#To validate using an external NLP library and printing entities
nlp = spacy.load('en_core_web_sm')
text = nlp(ClimateChange_DF.CleanedTweets[index])
displacy.render(text, style = 'ent', jupyter=True)

ClimateChange_DF.CleanedTweets[index]
